+++ components/button.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { TouchableOpacity, Text, StyleSheet } from 'react-native';
 
-
 interface buttonProps {
   title: string;     
   onPress: () => void;
