+++ package.json
@@ -11,25 +11,30 @@
   "dependencies": {
     "@expo-google-fonts/montserrat-alternates": "^0.2.3",
     "@expo/react-native-action-sheet": "^4.1.0",
-    "expo": "^51.0.38",
+    "@react-navigation/drawer": "^7.0.1",
+    "@rneui/base": "^4.0.0-rc.8",
+    "@rneui/themed": "^4.0.0-rc.8",
+    "expo": "^52.0.5",
     "expo-app-loading": "^2.1.1",
-    "expo-constants": "~15.4.5",
-    "expo-font": "^12.0.10",
-    "expo-image": "^1.13.0",
-    "expo-linking": "~6.2.2",
-    "expo-router": "~3.4.8",
-    "expo-status-bar": "~1.11.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.73.6",
+    "expo-constants": "~17.0.2",
+    "expo-font": "~13.0.1",
+    "expo-image": "~2.0.0",
+    "expo-linking": "~7.0.2",
+    "expo-router": "~4.0.5",
+    "expo-status-bar": "~2.0.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-native": "0.76.1",
     "react-native-action-sheet": "^2.2.0",
-    "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "~3.29.0",
+    "react-native-gesture-handler": "^2.20.2",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.0.0",
     "react-native-web": "~0.19.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "@types/react": "~18.2.45",
+    "@types/react": "~18.3.12",
     "typescript": "^5.1.3"
   },
   "private": true
